[features] File server - Upload File
