+++ file/src/main/kotlin/com/kohan/file/repository/FileRepository.kt
@@ -1,6 +1,6 @@
 package com.kohan.file.repository
 
-import com.kohan.file.collection.file.FileCollection
+import com.kohan.shared.collection.file.FileCollection
 import org.bson.types.ObjectId
 import org.springframework.data.mongodb.repository.MongoRepository
