+++ proto/src/main/proto/file/uploadFile.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package com.kohan.shared.armeria.file.v1;
+package com.kohan.proto.file.v1;
 
 message UploadFileInfo{
   string fileName = 2;
